requirements.txt
requests
schedule
python-dotenv
supabase
streamlit
